+++ tidy3d/components/structure.py
@@ -1,4 +1,6 @@
 """Defines Geometric objects with Medium properties."""
+from __future__ import annotations
+
 from typing import Union, Tuple, Optional
 import pydantic.v1 as pydantic
 import numpy as np
@@ -159,6 +161,15 @@ class Structure(AbstractStructure):
             _ = geom._normal_2dmaterial
 
         return val
+
+    def _compatible_with(self, other: Structure) -> bool:
+        """Whether these two structures are compatible."""
+        # note: if the first condition fails, the second won't get triggered
+        if not self.medium._compatible_with(other.medium) and self.geometry.intersects(
+            other.geometry
+        ):
+            return False
+        return True
 
     def eps_comp(self, row: Axis, col: Axis, frequency: float, coords: Coords) -> complex:
         """Single component of the complex-valued permittivity tensor as a function of frequency.